requirements.txt
amino.py
pyfiglet
colorama